Level cutout now uses bytes

+++ Mario-AI-Framework/src/engine/helper/SpriteType.java
@@ -52,7 +52,7 @@ public enum SpriteType {
         return new Enemy(visuals, xTile * 16 + 8, yTile * 16 + 15, dir, this);
     }
 
-    public static SpriteType getSpriteType(int value) {
+    public static SpriteType getSpriteType(byte value) {
         for (SpriteType spriteType : SpriteType.values()) {
             if (spriteType.value == value)
                 return spriteType;
